bikin blog laravel

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,17 @@ use Illuminate\Support\Facades\DB;
 use App\Post;
 class PostsController extends Controller
 {
+
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -48,6 +59,7 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
         return redirect('/posts')->with('success','Post created');
     }
@@ -75,6 +87,10 @@ class PostsController extends Controller
     {
         //
           $post = Post::find($id);
+            if (auth()->user()->id !== $post->user_id) {
+             
+            return redirect('/posts')->with('error','unauthrrized page');
+            }
             return view('posts.edit',['post'=>$post]);
 
     }
@@ -111,6 +127,10 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+         if (auth()->user()->id !== $post->user_id) {
+             
+            return redirect('/posts')->with('error','unauthrrized page');
+            }
         $post->delete();
              return redirect('/posts')->with('success','Post Removed');
